RAG性能调优/30-OCR+LLM/src/requirements.txt
openai
streamlit
pillow
llama-index
llama-index-core
llama-index-llms-openai-like
llama-index-llms-dashscope
llama-index-embeddings-dashscope
